Reduce precision

+++ contrast/survey/fortran/procedures.f90
@@ -5,10 +5,10 @@ contains
   subroutine linked_list(pos, ngrid, gridmin, gridmax, ll, lirst, rgrid)
     implicit none
     integer*8 :: i, ng, ipx, ipy, ipz
-    integer*8, intent(in) :: ngrid
-    real*8, intent(out) :: rgrid
-    real*8, intent(in) :: gridmin, gridmax
-    real*8, dimension(:,:), intent(in) :: pos
+    integer*4, intent(in) :: ngrid
+    real*4, intent(out) :: rgrid
+    real*4, intent(in) :: gridmin, gridmax
+    real*4, dimension(:,:), intent(in) :: pos
     integer*8, dimension(:,:,:), intent(out) :: lirst
     integer*8, dimension(:), intent(out) :: ll
 
@@ -45,8 +45,8 @@ contains
     integer*8 :: nrows, ncols
     character(len=500), intent(in) :: input_filename
     integer*8, intent(out) :: np
-    real*8, allocatable, dimension(:,:), intent(out) :: data
-    real*8, allocatable, dimension(:), intent(out) :: weight
+    real*4, allocatable, dimension(:,:), intent(out) :: data
+    real*4, allocatable, dimension(:), intent(out) :: weight
 
     open(20, file=input_filename, status='old', form='unformatted')
     read(20) nrows
@@ -66,11 +66,11 @@ contains
   subroutine binning(rmin, rmax, nbin, bin, bin_edges, rwidth)
     implicit none
 
-    integer*8 :: i
-    integer*8, intent(in) :: nbin
-    real*8, intent(in) :: rmin, rmax
-    real*8, intent(out) :: rwidth
-    real*8, allocatable, dimension(:), intent(out) :: bin, bin_edges
+    integer*4 :: i
+    integer*4, intent(in) :: nbin
+    real*4, intent(in) :: rmin, rmax
+    real*4, intent(out) :: rwidth
+    real*4, allocatable, dimension(:), intent(out) :: bin, bin_edges
 
     allocate(bin(nbin))
     allocate(bin_edges(nbin + 1))
